feat(starting_pose_matcher): expose create_main_widget() for embedded launch (part of #4998)
Subtask 5 / #4998 of EPIC #4993 — make the Motion-Match Preview
(``motion_target_preview``) launch as either a standalone QMainWindow
or an embedded QWidget inside the launcher host (tab / dock).

The original ``gui.py`` was the largest GUI in the codebase
(~3.1k lines). Split it across a thin shell plus a ``MainWidget`` and
three mixin modules so every file stays under the 1200-line budget:

- ``_gui_common.py`` — stylesheet, help text, ``LabelledControl``,
  shared helpers and constants.
- ``gui_render_mixin.py`` — ``_RenderMixin`` with axes setup, redraw
  + every ``_draw_*`` helper, residual computation, frame helpers.
- ``gui_builders_mixin.py`` — ``_BuildersMixin`` with every
  ``_build_<section>_box`` plus the engine-combo helpers.
- ``gui_session_mixin.py`` — ``_SessionMixin`` with offsets-export,
  full session save/load + ``_apply_session``.
- ``gui_main_widget.py`` — embeddable ``MainWidget(QWidget)`` that
  composes the mixins, owns ``__init__`` + ``_build_ui`` + the load
  / playback / transform / snap handlers, and adds a ``cleanup``
  hook for the embed lifecycle.
- ``gui.py`` — thin shell: ``StartingPoseMatcher(QMainWindow)`` now
  hosts ``MainWidget`` via ``setCentralWidget`` and forwards
  attribute access to it so legacy callers and the headless
  tooltip-coverage test keep working.

Wire-up:

- ``_embed_adapter._MotionMatchPreviewEmbedAdapter`` advertises
  ``tool_id="motion_target_preview"`` (the launcher tile id; the
  package keeps its historical ``starting_pose_matcher`` name) with
  ``min_size=(1024, 720)``, ``prefers_dock=False``, and a
  per-widget ``cleanup`` that stops the playback timer and tears
  down the live-view layers.
- ``__init__.py`` registers the adapter behind
  ``contextlib.suppress(ImportError)`` and a
  ``get_embeddable_tool`` idempotency guard, matching the
  data_explorer / model_explorer pattern.
- ``src/config/models.yaml`` adds ``default_launch: "tab"`` to the
  ``motion_target_preview`` block.
- ``scripts/config/file_size_budget.json`` drops the now-unneeded
  exception for ``gui.py`` (every new file is well under 1200
  lines).

Tests:

- New ``tests/ui/tools/starting_pose_matcher/test_embed_adapter.py``
  exercises protocol conformance, capabilities, ``create_main_widget``
  with and without a parent, idempotent cleanup, and the registry
  side-effect — all eight tests pass.
- Updated ``tests/unit/tools/starting_pose_matcher/test_engine_combo.py``
  to grab the engine-combo helpers off ``MainWidget`` instead of
  ``StartingPoseMatcher`` (they now live on the mixin).
- Existing matcher unit suite holds its pre-refactor pass/fail rate
  (218 passed / 49 pre-existing fleet-debt failures); the
  ``test_help_coverage`` matcher tooltip whitelist now passes
  because eager ``_live_body_target`` initialisation surfaces a
  long-latent missing tooltip on ``cb_lock_xy`` that the refactor
  adds.
