+++ src/tools/starting_pose_matcher/__init__.py
@@ -9,9 +9,23 @@ Public entry points:
 Public modules:
 
     .core                math, dataclasses, xlsx + skeleton + trajectory loaders
-    .gui                 PyQt6 QMainWindow
+    .gui                 PyQt6 standalone QMainWindow shell
+    .gui_main_widget     embeddable :class:`MainWidget` used by the launcher
     .skeleton_provider   pluggable source of model skeleton joints
+
+Importing this package registers the
+:class:`_MotionMatchPreviewEmbedAdapter` with the embeddable-tool
+registry so the launcher can host the tool as a tab or dock without
+spawning a separate process. Registration is guarded so reimports
+(test reloads) are a quiet no-op, and wrapped in
+``contextlib.suppress(ImportError)`` so headless contexts where PyQt6
+is unavailable still get a usable package. See Subtask 5 / #4998 of
+EPIC #4993.
 """
+
+from __future__ import annotations
+
+import contextlib
 
 from .core import (
     CM_TO_M,
@@ -35,6 +49,18 @@ from .core import (
     read_event_header,
     solve_shaft_rz_deg,
 )
+
+with contextlib.suppress(ImportError):
+    from src.shared.python.launcher_embed import (
+        get_embeddable_tool,
+        register_embeddable_tool,
+    )
+
+    from ._embed_adapter import _MotionMatchPreviewEmbedAdapter
+
+    _ADAPTER = _MotionMatchPreviewEmbedAdapter()
+    if get_embeddable_tool(_ADAPTER.tool_id) is None:
+        register_embeddable_tool(_ADAPTER)
 
 __all__ = [
     "CM_TO_M",
